Refactor import paths in Header and ConnectWalletButton components for consistency
